+++ src/reporter/query_1TNENA.py
@@ -3,6 +3,7 @@ from flask import request
 from reporter.reporter import _validate_query_params
 from translators.factory import translator_for
 import logging
+import warnings
 from .geo_query_handler import handle_geo_query
 from utils.jsondict import lookup_string_match
 import dateutil.parser
@@ -104,6 +105,7 @@ def query_1TNENA(entity_type=None,  # In Path
                                 from_date,
                                 to_date,)
         logging.getLogger(__name__).info("Query processed successfully")
+        logging.warn("usage of  id and type rather than entityId and entityType from version 0.9")
         return res
 
     r = {
@@ -158,4 +160,5 @@ def query_1TNENA_value(*args, **kwargs):
         res.pop('entityType', None)
         res['values'] = res['entities']
         res.pop('entities', None)
+    logging.warn("usage of  id and type rather than entityId and entityType from version 0.9")
     return res