+++ RELEASE_NOTES.md
@@ -11,6 +11,8 @@
 - Added exception handling and updated response where
   'AggrMethod cannot be applied' (#498)
 - Add workflow to check that `RELEASE_NOTES.md` is updated (#582)
+- Added a warning to use 'id' and 'type' from version 0.9 in all
+   query responses (#584)
 
 ### Bug fixes